move branch files to new directory
