+++ sympy/rules/__init__.py
@@ -29,3 +29,4 @@ from rl import rm_id, unpack, flatten, sort, glom, distribute
 from util import new
 from strat import (canon, condition, debug, typed, chain, null_safe, do_one,
         exhaust)
+import branch
