+++ sympy/rules/branch/__init__.py
@@ -1,0 +1,2 @@
+import traverse
+from strat import (canon, condition, debug, multiplex, exhaust, notempty)
