+++ sympy/rules/branch/strat_pure.py
