+++ sympy/rules/branch/tests/test_strat_pure.py
@@ -1,4 +1,4 @@
-from sympy.rules.branch_strat_pure import (exhaust, debug, multiplex,
+from sympy.rules.branch.strat_pure import (exhaust, debug, multiplex,
         condition, notempty)
